+++ simple_http_request/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:http/http.dart' as http;
-import 'dart:convert';
+import 'package:simple_http_request/ui/album_page.dart';
 
 void main() => runApp(MyApp());
 
@@ -13,82 +12,6 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
       ),
       home: AlbumPage(),
-    );
-  }
-}
-
-class Album {
-  final int userId;
-  final int id;
-  final String title;
-
-  Album({this.userId, this.id, this.title});
-
-  factory Album.fromJson(Map<String, dynamic> json) {
-    return Album(
-      userId: json['userId'],
-      id: json['id'],
-      title: json['title'],
-    );
-  }
-}
-
-class AlbumPage extends StatefulWidget {
-  @override
-  _AlbumPageState createState() => _AlbumPageState();
-}
-
-class _AlbumPageState extends State<AlbumPage> {
-  Future<Album> _futureAlbum;
-
-  @override
-  void initState() {
-    super.initState();
-    _futureAlbum = fetchAlbum();
-  }
-
-  Future<Album> fetchAlbum() async {
-    final response =
-        await http.get('https://jsonplaceholder.typicode.com/albums/1');
-
-    if (response.statusCode == 200) {
-      return Album.fromJson(json.decode(response.body));
-    } else {
-      throw Exception('Failed to load album');
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Fetch Data Example'),
-      ),
-      body: Center(
-        child: FutureBuilder<Album>(
-          future: _futureAlbum,
-          builder: (context, snapshot) {
-            var state = snapshot.connectionState;
-            switch (state) {
-              case ConnectionState.none:
-              case ConnectionState.active:
-              case ConnectionState.waiting:
-                return Center(child: CircularProgressIndicator());
-              case ConnectionState.done:
-                if (snapshot.hasData) {
-                  return Text(snapshot.data.title);
-                } else if (snapshot.hasError) {
-                  return Center(child: Text("${snapshot.error}"));
-                } else {
-                  return Text('');
-                }
-                break;
-              default:
-                return Text('');
-            }
-          },
-        ),
-      ),
     );
   }
 }
