+++ simple_http_request/lib/ui/album_page.dart
@@ -1,0 +1,57 @@
+import 'dart:convert';
+import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
+import 'package:simple_http_request/model/album.dart';
+
+class AlbumPage extends StatefulWidget {
+  @override
+  _AlbumPageState createState() => _AlbumPageState();
+}
+
+class _AlbumPageState extends State<AlbumPage> {
+  Future<Album> _futureAlbum;
+
+  @override
+  void initState() {
+    super.initState();
+    _futureAlbum = fetchAlbum();
+  }
+
+  Future<Album> fetchAlbum() async {
+    final response =
+        await http.get('https://jsonplaceholder.typicode.com/albums/1');
+    if (response.statusCode == 200) {
+      return Album.fromJson(json.decode(response.body));
+    } else {
+      throw Exception('Failed to load album');
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Fetch Data Example'),
+      ),
+      body: Center(
+        child: FutureBuilder<Album>(
+          future: _futureAlbum,
+          builder: (context, snapshot) {
+            var state = snapshot.connectionState;
+            if (state != ConnectionState.done) {
+              return Center(child: CircularProgressIndicator());
+            } else {
+              if (snapshot.hasData) {
+                return Text(snapshot.data.title);
+              } else if (snapshot.hasError) {
+                return Center(child: Text("${snapshot.error}"));
+              } else {
+                return Text('');
+              }
+            }
+          },
+        ),
+      ),
+    );
+  }
+}
